Move CheckboxGroup files in Checkbox folder
