+++ src/components/Checkbox/CheckboxGroup.scss
