+++ src/components/styles.scss
@@ -19,7 +19,7 @@
 @import './Tabs/Tabs';
 
 @import './Checkbox/Checkbox';
-@import './CheckboxGroup/CheckboxGroup';
+@import './Checkbox/CheckboxGroup';
 @import './DatePicker/DatePicker';
 @import './InputPhone/InputPhone';
 @import './InputSearch/InputSearch';
